Move PrimeErrorBoundary up the component tree to avoid double USA banner
