+++ frontend/src/app/PrimeErrorBoundary.tsx
@@ -1,0 +1,42 @@
+import React, { ErrorInfo } from "react";
+
+import ErrorPage from "./commonComponents/ErrorPage";
+import { getAppInsights } from "./TelemetryService";
+
+interface ErrorBoundaryState {
+  hasError: boolean;
+}
+
+export default class PrimeErrorBoundary extends React.Component<
+  {},
+  ErrorBoundaryState
+> {
+  constructor(props: any) {
+    super(props);
+    this.state = { hasError: false };
+  }
+
+  static getDerivedStateFromError(error: Error) {
+    // Update state so the next render will show the fallback UI.
+    return { hasError: true };
+  }
+
+  componentDidCatch(error: Error, info: ErrorInfo) {
+    const appInsights = getAppInsights();
+    if (appInsights) {
+      appInsights.trackException({
+        exception: new Error("PrimeErrorBoundary: " + error),
+        severityLevel: 3,
+        properties: { componentStack: info.componentStack },
+      });
+    }
+  }
+
+  render() {
+    if (this.state.hasError) {
+      return <ErrorPage />;
+    }
+
+    return this.props.children;
+  }
+}
