+++ frontend/src/app/signUp/SignUpApp.tsx
@@ -1,6 +1,5 @@
 import { Route, RouteComponentProps } from "react-router-dom";
 
-import PrimeErrorBoundary from "../PrimeErrorBoundary";
 import Page from "../commonComponents/Page/Page";
 
 import Consent from "./IdentityVerification/Consent";
@@ -8,16 +7,10 @@ import OrganizationForm from "./Organization/OrganizationForm";
 
 const SignUpApp: React.FC<RouteComponentProps> = ({ match: { path } }) => {
   return (
-    <PrimeErrorBoundary>
-      <Page>
-        <Route path={`${path}`} exact component={OrganizationForm} />
-        <Route
-          path={`${path}/identity-verification`}
-          exact
-          component={Consent}
-        />
-      </Page>
-    </PrimeErrorBoundary>
+    <Page>
+      <Route path={`${path}`} exact component={OrganizationForm} />
+      <Route path={`${path}/identity-verification`} exact component={Consent} />
+    </Page>
   );
 };
