+++ frontend/src/patientApp/PatientApp.tsx
@@ -2,7 +2,6 @@ import { FunctionComponent, useEffect } from "react";
 import { useDispatch, connect, useSelector } from "react-redux";
 import { Route, Switch, BrowserRouter as Router } from "react-router-dom";
 
-import PrimeErrorBoundary from "../app/PrimeErrorBoundary";
 import Page from "../app/commonComponents/Page/Page";
 import { setInitialState } from "../app/store";
 import { getPatientLinkIdFromUrl } from "../app/utils/url";
@@ -49,45 +48,43 @@ const PatientApp = () => {
   });
 
   return (
-    <PrimeErrorBoundary>
-      <Page>
-        <PatientHeader />
-        <PatientLinkURL404Wrapper plid={plid}>
-          <Router basename={`${process.env.PUBLIC_URL}/pxp`}>
-            <Switch>
-              <Route path="/" exact component={TermsOfService} />
-              <Route path="/terms-of-service" component={TermsOfService} />
-              <Route path="/birth-date-confirmation" component={DOB} />
-              <GuardedRoute
-                auth={auth}
-                path="/patient-info-confirm"
-                component={PatientProfileContainer}
-              />
-              <GuardedRoute
-                auth={auth}
-                path="/patient-info-edit"
-                component={PatientFormContainer}
-              />
-              <GuardedRoute
-                auth={auth}
-                path="/patient-info-symptoms"
-                component={AoEPatientFormContainer}
-              />
-              <GuardedRoute
-                auth={auth}
-                path="/success"
-                component={PatientLanding}
-              />
-              <GuardedRoute
-                auth={auth}
-                path="/test-result"
-                component={TestResult}
-              />
-            </Switch>
-          </Router>
-        </PatientLinkURL404Wrapper>
-      </Page>
-    </PrimeErrorBoundary>
+    <Page>
+      <PatientHeader />
+      <PatientLinkURL404Wrapper plid={plid}>
+        <Router basename={`${process.env.PUBLIC_URL}/pxp`}>
+          <Switch>
+            <Route path="/" exact component={TermsOfService} />
+            <Route path="/terms-of-service" component={TermsOfService} />
+            <Route path="/birth-date-confirmation" component={DOB} />
+            <GuardedRoute
+              auth={auth}
+              path="/patient-info-confirm"
+              component={PatientProfileContainer}
+            />
+            <GuardedRoute
+              auth={auth}
+              path="/patient-info-edit"
+              component={PatientFormContainer}
+            />
+            <GuardedRoute
+              auth={auth}
+              path="/patient-info-symptoms"
+              component={AoEPatientFormContainer}
+            />
+            <GuardedRoute
+              auth={auth}
+              path="/success"
+              component={PatientLanding}
+            />
+            <GuardedRoute
+              auth={auth}
+              path="/test-result"
+              component={TestResult}
+            />
+          </Switch>
+        </Router>
+      </PatientLinkURL404Wrapper>
+    </Page>
   );
 };
